move numpy_to_vanilla
